+++ src/main/java/ladder/formatter/PlayerNameFormatter.java
@@ -1,15 +1,17 @@
-package ladder.view;
+package ladder.formatter;
 
-import ladder.domain.Prize;
+import ladder.domain.Player;
 
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class PrizeNameFormatter implements NameFormatter<Prize> {
+public class PlayerNameFormatter implements NameFormatter<Player>{
+//    private static final int SPACE_FOR_NAME = 5;
+//    private static final String BLANK_TO_FILL_THE_NAME_SPACE = " ";
 
     @Override
-    public String nameFormat(Prize prize) {
-        String name = prize.getPrize();
+    public String nameFormat(Player player) {
+        String name = player.getName();
         int spaceForBlank = SPACE_FOR_NAME - name.length();
 
         return IntStream.rangeClosed(0, spaceForBlank)
