+++ src/main/java/ladder/formatter/PrizeNameFormatter.java
@@ -1,17 +1,15 @@
-package ladder.view;
+package ladder.formatter;
 
-import ladder.domain.Player;
+import ladder.domain.Prize;
 
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class PlayerNameFormatter implements NameFormatter<Player>{
-//    private static final int SPACE_FOR_NAME = 5;
-//    private static final String BLANK_TO_FILL_THE_NAME_SPACE = " ";
+public class PrizeNameFormatter implements NameFormatter<Prize> {
 
     @Override
-    public String nameFormat(Player player) {
-        String name = player.getName();
+    public String nameFormat(Prize prize) {
+        String name = prize.getPrize();
         int spaceForBlank = SPACE_FOR_NAME - name.length();
 
         return IntStream.rangeClosed(0, spaceForBlank)
